+++ packages/sources/cfbenchmarks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chainlink/cfbenchmarks-adapter
 
+## 1.7.0
+
+### Minor Changes
+
+- 6f605b8be: Added custom data validation & custom retry config for cfbenchmarks BIRC
+
 ## 1.6.1
 
 ### Patch Changes